Move extractor.
